Mock Server support for member list (#2860)
