+++ TestTools/StreamChatTestMockServer/MockServer/ChannelResponses.swift
@@ -395,10 +395,6 @@ public extension StreamMockServer {
         return channelDetails
     }
 
-    private func findChannelById(_ id: String) -> [String: Any]? {
-        try? XCTUnwrap(waitForChannelWithId(id))
-    }
-
     func getFirstChannelId() -> String {
         let endTime = TestData.waitingEndTime
         while channelList.isEmpty && endTime > TestData.currentTimeInterval {}
@@ -444,19 +440,6 @@ public extension StreamMockServer {
             channels.append(channel)
             channelList[JSONKey.channels] = channels
         }
-    }
-
-    private func waitForChannelWithId(_ id: String) -> [String: Any]? {
-        let endTime = TestData.waitingEndTime
-        var newChannelList: [[String: Any]] = []
-        while newChannelList.isEmpty && endTime > TestData.currentTimeInterval {
-            guard let channels = channelList[JSONKey.channels] as? [[String: Any]] else { return nil }
-            newChannelList = channels.filter {
-                let channel = $0[JSONKey.channel] as? [String: Any]
-                return id == channel?[channelKey.id.rawValue] as? String
-            }
-        }
-        return newChannelList.first
     }
 
     private func channelTruncation(_ request: HttpRequest) -> HttpResponse? {
